refactor move folder specification to infra
