+++ src/main/java/br/com/mili/milibackend/fornecedor/infra/specification/GfdTipoDocumentoSpecification.java
@@ -1,4 +1,4 @@
-package br.com.mili.milibackend.fornecedor.domain.specification;
+package br.com.mili.milibackend.fornecedor.infra.specification;
 
 import br.com.mili.milibackend.fornecedor.domain.entity.GfdTipoDocumento;
 import br.com.mili.milibackend.fornecedor.domain.entity.GfdTipoDocumentoTipoEnum;
